server/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.0.1
Django==4.1.7
django-channels==0.7.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
environ==1.0
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
